--- web/src/components/LoginModal/LoginModal.stories.tsx
@@ -1,16 +1,0 @@
-import { Meta, StoryObj } from '@storybook/react';
-
-import LoginModal from './LoginModal';
-
-export default {
-  title: 'Components/LoginModal',
-  component: LoginModal,
-} as Meta;
-
-const Template: StoryObj<typeof LoginModal> = {
-  args: {
-    login: () => console.log('Logged in'),
-  },
-};
-
-export const Default = Template;
